added new utils
* Refactored collection package into separate files
